suppress compiler warnings generated when compiling with single precision e.g. always cast DSL const variables to the datatype to suppress warnings about implicit casting from double literals to single

+++ src/core/analysis_grid_helpers.h
@@ -311,7 +311,7 @@ get_stencil_dims(const Stencil stencil)
 		{
 			for(int z = -NGHOST; z <= NGHOST; ++z)
 			{
-				if(stencils[stencil][x+NGHOST][y+NGHOST][z+NGHOST] != 0.0)
+				if(double(stencils[stencil][x+NGHOST][y+NGHOST][z+NGHOST]) != 0.0)
 				{
 					min_radius.x = std::min(min_radius.x,x);
 					min_radius.y = std::min(min_radius.y,y);
